refactor: isAdmin에 따른 페이지 분기 적용
